// src/app/model/user.ts
export class User {
    id: string;
    firstName: string;
    lastName: string;
    codeMosad: number;
    phone?: string;
    email?: string;
    password?: string;

    // constructor(id: string, firstName: string, lastName: string, codeMosad: number, phone?: string, email?: string,) {

    // }
}